pyproject.toml
[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[project]
name        = "sql-analytics-env"
version     = "0.1.0"
description = "SQL Analytics OpenEnv environment — BI question-answering via SQL"
readme      = "README.md"
license     = {text = "MIT"}
requires-python = ">=3.10"

authors = [
    {name = "SQL Analytics Env", email = "sql-analytics@example.com"}
]

keywords = [
    "sql", "database", "reinforcement-learning",
    "openenv", "business-intelligence", "ecommerce", "agent"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openenv-core>=0.1.0,<1.0.0",
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.29.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "openai>=1.30.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage      = "https://huggingface.co/spaces/your-username/sql-analytics-env"
Repository    = "https://github.com/your-username/sql-analytics-env"
Documentation = "https://github.com/your-username/sql-analytics-env#readme"

[project.scripts]
sql-analytics-server = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths   = ["."]
python_files = ["test_*.py"]
